+++ src/pipeline_tools/tools/project_creator/gitignore_templates.py
@@ -38,6 +38,7 @@ htmlcov/
 *.temp
 *.bak
 *.backup
+05_WORK/
 *_backup
 """